将Long类型的改为integer 在建表文件里面增加了一些表，和实体类对应 添加了游戏上传相关配置 将多余的findByEmail方法删去 在前端页面添加发送验证码和输入验证码的按钮

+++ stream_game/src/main/java/com/xtu/stream_game/entity/Transaction.java
@@ -13,7 +13,7 @@ import java.util.Date;
 public class  Transaction{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long transactionId;
+    private Integer transactionId;
     @ManyToOne
     @JoinColumn(name = "player_id")
     private Player player;
@@ -29,11 +29,11 @@ public class  Transaction{
     }
 
     // Getters and Setters
-    public Long getTransactionId() {
+    public Integer getTransactionId() {
         return transactionId;
     }
 
-    public void setTransactionId(Long transactionId) {
+    public void setTransactionId(Integer transactionId) {
         this.transactionId = transactionId;
     }
